# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DriveApp API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type DriverStatus = (typeof DriverStatus)[keyof typeof DriverStatus];

export const DriverStatus = {
  available: "available",
  busy: "busy",
  offline: "offline",
} as const;

export interface Driver {
  id: number;
  name: string;
  phone: string;
  email: string;
  licenseNumber: string;
  vehicleMake: string;
  vehicleModel: string;
  vehicleYear: number;
  vehiclePlate: string;
  vehicleColor: string;
  status: DriverStatus;
  rating: number;
  totalRides: number;
  createdAt: string;
}

export interface RegisterDriverRequest {
  name: string;
  phone: string;
  email: string;
  licenseNumber: string;
  vehicleMake: string;
  vehicleModel: string;
  vehicleYear: number;
  vehiclePlate: string;
  vehicleColor: string;
}

export type UpdateDriverStatusRequestStatus =
  (typeof UpdateDriverStatusRequestStatus)[keyof typeof UpdateDriverStatusRequestStatus];

export const UpdateDriverStatusRequestStatus = {
  available: "available",
  busy: "busy",
  offline: "offline",
} as const;

export interface UpdateDriverStatusRequest {
  status: UpdateDriverStatusRequestStatus;
}

export type RideStatus = (typeof RideStatus)[keyof typeof RideStatus];

export const RideStatus = {
  pending: "pending",
  accepted: "accepted",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Ride {
  id: number;
  clientName: string;
  clientPhone: string;
  pickupLocation: string;
  pickupLat: number;
  pickupLng: number;
  dropoffLocation: string;
  dropoffLat: number;
  dropoffLng: number;
  distanceKm: number;
  driverId?: number | null;
  driver?: Driver | null;
  status: RideStatus;
  estimatedFare: number;
  finalFare?: number | null;
  notes?: string | null;
  createdAt: string;
  acceptedAt?: string | null;
  startedAt?: string | null;
  completedAt?: string | null;
}

export interface CreateRideRequest {
  clientName: string;
  clientPhone: string;
  pickupLocation: string;
  pickupLat: number;
  pickupLng: number;
  dropoffLocation: string;
  dropoffLat: number;
  dropoffLng: number;
  notes?: string;
}

export type UpdateRideStatusRequestStatus =
  (typeof UpdateRideStatusRequestStatus)[keyof typeof UpdateRideStatusRequestStatus];

export const UpdateRideStatusRequestStatus = {
  accepted: "accepted",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface UpdateRideStatusRequest {
  status: UpdateRideStatusRequestStatus;
  driverId?: number;
}

export type BillStatus = (typeof BillStatus)[keyof typeof BillStatus];

export const BillStatus = {
  unpaid: "unpaid",
  paid: "paid",
} as const;

export interface Bill {
  id: number;
  rideId: number;
  ride?: Ride | null;
  driverId?: number | null;
  clientName: string;
  pickupLocation: string;
  dropoffLocation: string;
  distanceKm: number;
  baseFare: number;
  distanceFare: number;
  totalFare: number;
  currency: string;
  status: BillStatus;
  createdAt: string;
}

export type ListDriversParams = {
  status?: ListDriversStatus;
};

export type ListDriversStatus =
  (typeof ListDriversStatus)[keyof typeof ListDriversStatus];

export const ListDriversStatus = {
  available: "available",
  busy: "busy",
  offline: "offline",
} as const;

export type ListRidesParams = {
  driverId?: number;
  status?: ListRidesStatus;
};

export type ListRidesStatus =
  (typeof ListRidesStatus)[keyof typeof ListRidesStatus];

export const ListRidesStatus = {
  pending: "pending",
  accepted: "accepted",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export type ListBillsParams = {
  driverId?: number;
};