# lib/api-zod/src/generated/types/bill.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DriveApp API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BillStatus } from "./billStatus";
import type { Ride } from "./ride";

export interface Bill {
  id: number;
  rideId: number;
  ride?: Ride | null;
  driverId?: number | null;
  clientName: string;
  pickupLocation: string;
  dropoffLocation: string;
  distanceKm: number;
  baseFare: number;
  distanceFare: number;
  totalFare: number;
  currency: string;
  status: BillStatus;
  createdAt: Date;
}