# lib/api-zod/src/generated/types/createRideRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DriveApp API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateRideRequest {
  clientName: string;
  clientPhone: string;
  pickupLocation: string;
  pickupLat: number;
  pickupLng: number;
  dropoffLocation: string;
  dropoffLat: number;
  dropoffLng: number;
  notes?: string;
}