# lib/api-zod/src/generated/types/registerDriverRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DriveApp API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterDriverRequest {
  name: string;
  phone: string;
  email: string;
  licenseNumber: string;
  vehicleMake: string;
  vehicleModel: string;
  vehicleYear: number;
  vehiclePlate: string;
  vehicleColor: string;
}