# lib/api-zod/src/generated/types/ride.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DriveApp API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Driver } from "./driver";
import type { RideStatus } from "./rideStatus";

export interface Ride {
  id: number;
  clientName: string;
  clientPhone: string;
  pickupLocation: string;
  pickupLat: number;
  pickupLng: number;
  dropoffLocation: string;
  dropoffLat: number;
  dropoffLng: number;
  distanceKm: number;
  driverId?: number | null;
  driver?: Driver | null;
  status: RideStatus;
  estimatedFare: number;
  finalFare?: number | null;
  notes?: string | null;
  createdAt: Date;
  acceptedAt?: Date | null;
  startedAt?: Date | null;
  completedAt?: Date | null;
}