# lib/api-zod/src/generated/types/updateDriverStatusRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DriveApp API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateDriverStatusRequestStatus } from "./updateDriverStatusRequestStatus";

export interface UpdateDriverStatusRequest {
  status: UpdateDriverStatusRequestStatus;
}